requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
djangorestframework == 3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
ruff==0.6.7
django-filter==24.3
django-cors-headers==4.4.0
django-silk==5.2.0
